Text can be folded by Ctrl+Clicking on tags like #[x] #{xx} or similar.
Note: Sometimes, we may use parentheses, brackets, or other symmetrical characters to improve visibility, for aesthetic reasons, as in:

This paragraph... blah, blah... (#Critical )

Note the space before the closing parenthesis. Without it, KeyNote will attempt to create a new tag named "Critical)".
This is because any character is valid within a tag, except: [' ', ':', ',', #13, #9, '#']