Export: It's now possible to restrict the nodes considered or even the text fragments exported using Tag criteria.
The same controls and features available in Find All have been added to the Export form to allow you to exclude
nodes that don't meet the tag criteria. These new controls allow you to export only the fragments of notes that
do meet the criteria.
For example, we could exclude fragments tagged with #private from the export.