Fixed: KNTLinks were not being converted in MergeFile operations (2.0+), as imported notes were incorrectly identified as HTML.
TNoteEntry.IsHTML returned the opposite (Not IsHTML). The TKntFile.ConvertKNTLinksToNewFormatInNotes method used this
property to ignore note entries whose content did not include KNT links. Because of this bug, the method did not convert any links.
As a result, an internal link in the imported file could point to the wrong node, as the GID of the imported nodes was converted,
but in this case, the included links were not.