Reverted the change for solving issue #830
It seems that it is the cause of this issue #854, besides issue #858

Although I have solved issue #858 while maintaining the change necessary to fix this issue #830 (Application.MainFormOnTaskbar := True), the impact in other aspects, like this issue #854, are more complex, specially by the influence on other controls like VirtualTreeView and RichText.
Only when all other issues are identified and solved I will try to reincorporate it.