Replacement operations will ignore matches located entirely or partially in the hidden part of hyperlinks (#871).
Replacements will be possible on the visible part, but not on the associated URL, which remains hidden.
In the case of URLs not built on hyperlinks, where the visible part matches the URL and the editor interprets it as a hyperlink (if the "Detect and highlight URLs in editor" option is enabled), the replacement operation is allowed.