Fixed: Possible exception "Failed to Load Stream" when working on newly created note/node, before it is persisted to the model
In a newly created node, before it is persisted in the model (for example, when selecting another node from the tree or when saving the file), an exception
"Failed to Load Stream" could occur on some situations.
For example, if we paste text in a new note and immediately select several lines and press Shift+TAB (to tab multiple lines, and reduce indentation)
the exception could raise. However, if we perform an RTF operation, such as making the text bold, before pasting (or typing the text), the Shift+TAB operation will work correctly.