Fixed: All possible exceptions that may occur in ApplicationEventsIdle are discarded.
These would be very rare exceptions, due to the possible assignment of the active paragraph's RTL property, or an attempt to display the tag selector.
However, if they occur, they could be very dangerous, as they would probably be repeated continuously, making it impossible to continue and save (or save as).