Folding/Unfolding can now be used with tables (the table structure, with its borders, will also be hidden.)
The Folding/Unfolding mechanism in KeyNote uses the \v and \v0 RTF tags to hide text. It's considerably more complicated than simply adding these tags to the beginning and end of the text to be collapsed, as there are several restrictions that must be met, such as the fact that nesting \v commands is not allowed, nor is hiding images or links, etc. To avoid these problems, KeyNote must perform some conversions before hiding the text (leaving only the necessary excerpt visible).
With the implementation available up to KeyNote version 2.1.0, if you try to fold text that includes tables, you may notice that the text is hidden, but the tables remain visible, albeit empty.

Nested Cells
In addition to the usual formatting indicated for tables, RTF allows for the inclusion of cells within other cells, using commands such as \nestcell.
Managing tables with nested cells would unnecessarily complicate the entire processing process and increase the possibility of errors that could result in malformed RTF code and therefore potentially lead to the loss of information. For this reason, KeyNote NF will *NOT* initially support tables with nested cells. When applying folding, if it detects the \nestcell command, it will issue a warning and prevent these changes from being applied.

Although folding is not supported on tables with nested cells, it is possible to apply folding in text fragments with tables that include images, links, or even other nested blocks as the content of one or more cells. These nested blocks can, in turn, contain other tables, images, text, and so on.