Fixed: The closing symbol for a folded block, ...[ ], will now always display in the same font and size.
Until now, if the collapsed text includes fonts like Arial, Courier New, and others at the end, the character used is not displayed as a little square but as a vertical rectangle. Its size also varies depending on the text content, which can make it very large or very small.
It will now always display in the same size, corresponding to Tahoma 8.