Fixed: The Tree > Set Focus > "Toggle Focus Editor / Tree Panel" menu entry had the keyboard shortcut Ctrl+Esc. It should be Ctrl+º. (#887)
It was assigned the Ctrl+Esc keyboard shortcut, which is used by Windows to open the Start Menu. However, this was never a problem since internally,
KeyNote also uses the Ctrl+\ (Ctrl+º) shortcut, as indicated in the help.
From the Delphi IDE, it is not possible to assign Ctrl+º (but Ctrl+\ can). Perhaps that could be the source of the error. It is possible to assign
it through Tools -> Customize Keyboard... . By doing so, I was able to check the code associated with the Ctrl+º shortcut, and I associated it by
directly editing kn_Main.dfm.