Folding: Improved selection (with Ctrl+Click) of opening token when set before a carriage return
We might have something like this:
<<
Line 1
Line 2
>>

Where << is an opening token. If we Ctrl+Click slightly to the right of <<, we'll select the carriage return. To make it easier to use
tokens on this situations, in these cases we'll check if there's a word to the left and if it's a token. If it is, we'll use it.