Adjustment to previous commit "* Fixed: The closing symbol for a folded block, ...[ ], will now always display in the same font and size."
Instead of:
      ReplaceWith:= '{\f0\fs16' + KNT_RTF_END_FOLDED + '}\par}'
      ReplaceWith:= '{\f0\fs16' + KNT_RTF_END_FOLDED + '}}';

It will be:
      ReplaceWith:= '\v0{\f0\fs20' + KNT_RTF_END_FOLDED_WITHOUT_v0 + '}\par}'
      ReplaceWith:= '\v0{\f0\fs20' + KNT_RTF_END_FOLDED_WITHOUT_v0 + '}}';

This minor change ensures that the folding correctly respects the final line break.
For example, when folding the three bulleted lines:

--------
* One
* Two
* Three
--------

It should lead to:

--------
* One...[]
--------

But it was getting:
--------
* One...[]--------

The adjustment now resolves the issue.
I also took the opportunity to slightly increase the size of the [] character, from Tahoma 8 to Tahoma 10.