Fixed: Scroll bar position could be lost when focus was returned to KeyNote by clicking in the Editor (#720)
The problem is that when you click on the KeyNote editor window, it ignores the current scroll position and switches to displaying the position of the caret.
That is, if we have the caret in a position and we display the content of previous or subsequent lines using the scroll bar or Ctrl+Up/Down (depending on the configuration associated with that shortcut), when you switch to another application and then return to KeyNote, clicking on the editor no longer respects what the editor was displaying, but instead switches to displaying the position of the caret.

This problem doesn't occur in Windows 10
It only occurs in Windows 11 and only if the app isn't restored from the minimized state at some point. Additionally, setting the focus to the application by clicking on the tree panel or the resource panel will prevent or resolve this issue from that point forward in the session.

As soon as the application is minimized, by pressing ESC (if configured so), clicking the corresponding button in the application's title bar, or clicking the application's icon in the taskbar, the error will no longer occur once it is restored during the session.
Again, it doesn't matter how it is restored: by clicking the icon in the system tray, using the Hot Key, or Alt+TAB -- "Minimize to system tray" is not selected. It is possible to close and open the same or a different file; the problem no longer occurs.

Therefore, this problem shouldn't occur for users who have the "Start program minimized" option enabled or who use the ESC button to minimize ("On Escape key" = "Minimize main window").
I personally have both of those options enabled and am used to exiting the application with ESC, minimizing it, and returning to it by pressing the activation hotkey.

This issue has some relation to the problem reported in issue #720, already solved with commit 13ba02dab (* Fixed: Scrollbar position could be forgotten when returning to app with Hotkey (Ctrl+Shift+...)