* Fixed - IMPORTANT!: It was possible for new nodes to be created with an incorrect RTF/Plain state (See details).
If the option "Inherit properties from active node" (in General Settings | Tree Panel) is checked, the new nodes were created with the same
RTF/Plain state as the inherited node, independently of the "Default plain text only" option of the folder.
It shouldn't, as it is confused.
This behaviour, now fixed, it is not dangerous, only confusing. But there is other situation that it could lead to lose data, and so it is IMPORTANT:

When a node is created copying from another node, using Copy (or Cut) and Paste or drag and drop, and the source node is RTF
and "Default plain text only" is True in the target folder (or node is Plain and the option if False), then the content of the node will be
incorrectly saved.
Depending on it is created as RTF or Plain, and on it is the last created node before saving the file, it can cause the whole first folder to be hidden (*)
or only the next created node (after the problematic node) to be hidden (*)

Depending on whether it was created as RTF or plain text, and whether it was the last node created before saving the file, it could cause the entire
first folder to be hidden (*) or only the next node created (after the problematic node) to be hidden (*)

(*) The saved .knt still contains all the information hidden and it is easy to recover, BUT if the file is saved again (while the node or folder is hidden),
then that information will be lost, and it will necessary to use a backup to restore that information.

To correct this problem if you were affected you only have to remove an incorrect semicolon (if .knt is saved in a native format --not compressed nor encrypted):

A node will be hidden in this situation:
%*
GI=25
ND=IT IS RTF!!...
LM=0811251147
%.
DC=0811251147
NS=0002
%>
;{\rtf1\ansi\ansicpg1252\deff0\nouicompat\deflang3082{\fonttbl{\f0\fnil\fcharset0 Tahoma;}}
;{\*\generator Riched20 10.0.22621}\viewkind4\uc1
;\pard\f0\fs20 Hello WORLD!\par
;}
;%*                              <--- Remove ";"
GI=4
ND=Node HIDDEN...°

The first whole folder will be hidden in this situation:

%*
GI=25
ND=IT IS RTF!!...
LM=0811251147
%.
DC=0811251147
NS=0002
%>
;{\rtf1\ansi\ansicpg1252\deff0\nouicompat\deflang3082{\fonttbl{\f0\fnil\fcharset0 Tahoma;}}
;{\*\generator Riched20 10.0.22621}\viewkind4\uc1
;\pard\f0\fs20 Hello WORLD!\par
;}
;%+                               <--- Remove ";"
NN=Folder HIDDEN

If the node was Plain copied to a RTF default folder, it could cause no problem depending on the content of the note.
If nothing incorrect happened then, when the node is modified it will correctly be saved to a normal RTF node.
But if it was causing problems, then it is safer to save the content of the note to other file, and remove the content in the node:

%*
GI=28
ND=IT'S PLAIN!
LM=0811251255
%.
DC=0811251255
%:
Hello World
Line 2
%+
NN=Example

===>

%*
GI=28
ND=IT'S PLAIN! - TO BE REMOVED
LM=0811251255
%.
DC=0811251255
%:
%+
NN=Example