* File|Export: Improvements when using 'Consider extracts from notes'
The removal of fragments to be excluded has been improved, taking into account the With and Without options (together with 'Consider extracts from notes'),
so that they do not alter line breaks where they are located, and therefore any bullet points.
Now, if we have auxiliary tagged text inside the bullet list, we can exclude them in the export, while maintaining the bullet style:

* Requirement 1 to be validated from user #Tec: Internal or technical info that we don't waht to show to the final user##
* Requirement 2
* Requirement 3

With [Without: Tec] the export will be:

* Requirement 1 to be validated from user
* Requirement 2
* Requirement 3

Note: "Remove" check is only needed, optionally, to remove the tags from the fragments considered (not excluded) after processing.
The fragments identified and excluded because of the "Without" field will be removed in the processing, including its tag. So, in the following text
<<
Line 1
Line 2 #Tec: Bla,bla ##Final
Line 3
>>

If we indicate [Without: Tec] the exported text will be:
<<
Line 1
Line 2 Final
Line 3
>>

alhough "Remove" is unchecked.

* Also fixed possible exception if active editor was showing the tag selector when opening File|Export and entering in With or Without fields