Getting more information to analYze possible exceptions (with JCL (JEDI Code Library))
The module's base address is also displayed in exception message pop-ups to allow the location of the source code that generates
the exception using the "AddrLookup" utility (post-mortem address analyzer).

Used JCL (JEDI Code Library) to log the stack trace in exceptions (for use in executables generated in debug mode)

For executables that include debugging functionality, the -debug parameter should be used
You can optionally define the detail of the log with a number: example -debug4

C:\APL\KeyNoteNF\keynote.exe -debug4

The larger the number, the greater the detail (and log size). The maximum value used so far is -debug5, where a small initial
fragment of the notes loaded into the editor will be sent to the log.