Fixed: KNT links in previous format (file:///*) in Merged notes must be converted to new format based on GIDs
Applies to:  Tools|Merge Folders...

This is necessary because the folder IDs they are based on may have changed after the merge.
Instead of simply adapting the folder ID, these links are converted to the new format.
(Moving a note to another folder does not break any Knt link based in the new format)