Fixed: When unfolding a block with images, these could remain hidden, if it is contiguous with another block or hyperlink
Added code to protect the unfolding/extracting from what appears to be a bug in the RichEdit control:
If a collapsed text block contains an image and is immediately followed (or separated by a single line break) by another collapsed
block or a regular hyperlink, the image is hidden. This is because the `\v0` tag, which delimits the end of the hidden marker with
the image ID and should be positioned just before the image, is shifted to the right of the image itself, thereby hiding it.
This doesn't happen in the RTF returned by PrepareRTFtoBeExpanded, where the image displays correctly. However, when inserted right
next to these elements, it undergoes this transformation.
To avoid this, if this situation is detected, some auxiliary characters are inserted just after the end of the text to be expanded,
thus preventing this bug. Once the replacement is made, these characters are searched for and removed.

      <FoldeBlock1>
      <FoldeBlock2>

      <FoldeBlock1>
      <Hyperlink>

      <FoldeBlock1><Hyperlink>

      <FoldeBlock1><FoldeBlock2>