New: Insert | Line
A new command has been added that allows you to insert a line into the active editor, provided the editor is not a plain text editor.

An option has been added to Configuration Options (F5) -> Rich Text Editor: "Insert | Line: Limit to editor width on creation."
If selected, the line will be created with the current editor width; otherwise, it will be created with a very large width, extending
beyond any possible page width. In the first case (option selected), the right end of the line will be accessible, and you can drag
and drag the cursor over it to change the width.

You can insert the line using the menu entry: Insert | Line.
Clicking this entry while holding down the Shift key will insert a printable line; otherwise, the inserted line will be visible on
screen but will not appear when printed.
The default keyboard shortcuts are:
Ctrl+Numpad -		 Insert line
Ctrl+Shift+Numpad -	 Insert printable line

As with most commands, it's possible to change the keyboard shortcut for Insert Line. Pressing Shift at the same time will insert a printable line.

The two new commands (Insert Line and Insert Printable Line) can be repeated (Edit|Repeat...) and used within macros.

NOTE:
The RichEdit control does not allow to insert lines exclusively; it does not recognize the specific RTF command for that purpose.
The line inserted here is constructed using a cell, adjusted to appear as a line.

This has two limitations:
- The width is not dynamic; it is fixed and is set when the line is created.
- To achieve the line effect, a space is included within the line, with a font size of 1. This means that if you apply a font size change
  to a selection that includes a line, the cell will become visible.

The fact that the font size used within the cell is almost zero (it appears as 0 in the font size dropdown) offers a simple way to
delete a line: we can press the cursor keys until the font size displays 0. If we then press the down arrow key while holding down
the Shift key, we will select the entire line. We can then delete it (Delete).

The RTF code used to represent the line is:

Simple line (not printable):
        {\rtf1\ansi\trowd\cellx99999 \pard\intbl\fs1\cell\row \pard}

Printable line:
		{\rtf1\ansi{\colortbl ;\red255\green255\blue255;}
		\trowd\trgaph10\trpaddl10\trpaddr10\trpaddfl3\trpaddfr3
		\clbrdrt\brdrw10
		\clbrdrb\brdrw1\brdrcf1
		\clbrdrl\brdrw1\brdrcf1
		\clbrdrr\brdrw1\brdrcf1
		\cellx99999 \pard\intbl\fs1\cell\row \pard}