New: Insert | Table...
A new command has been added that allows you to insert a basic table into the active editor, provided the editor is not a plain text editor.
You will then be asked to specify the number of rows and columns in the table in the format "R,C".

You can insert the table using the menu entry: Insert | Table...
Clicking this entry while holding down the Shift key will insert a table whose lines will be printable; otherwise, the lines of the table
will be visible on screen but will not appear when printed.

The two new commands (Insert Table and Insert Printable Table) can be repeated (Edit|Repeat...) and used within macros.
When used in a macro, the command considers the last defined number of rows and columns in the session. If not yet defined, it will use 2 x 2.